requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.5.2
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.1.1
cycler==0.12.1
db-dtypes==1.4.2
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.16.1
fonttools==4.57.0
fsspec==2025.3.0
google-api-core==2.25.0
google-auth==2.40.3
google-cloud-bigquery==3.30.0
google-cloud-core==2.4.3
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gradio==4.44.1
gradio-client==1.3.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.4
idna==3.10
importlib-resources==6.4.5
iniconfig==2.1.0
jinja2==3.1.6
jiter==0.9.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
numpy==1.24.4
openai==1.85.0
openai-agents==0.0.1
orjson==3.10.15
packaging==25.0
pandas==2.0.3
pillow==10.4.0
pluggy==1.5.0
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.10.6
pydantic-core==2.27.2
pydub==0.25.1
pygments==2.19.1
pyparsing==3.1.4
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
rsa==4.9.1
ruff==0.11.13
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.44.0
tomli==2.2.1
tomlkit==0.12.0
tqdm==4.67.1
typer==0.16.0
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
uvicorn==0.33.0
websockets==12.0
zipp==3.20.2
